requirements.txt
requests
flask
python-dotenv
supabase
